+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [5.5.0] - 2019-04-30
+
 ### Added
 
 - [Box](/#/Layout?id=box) component.